feat: clear dir
