--- src/component/Topic/index.less
@@ -1,17 +1,0 @@
-.list {
-  :global {
-    .ant-card {
-      padding: 0;
-      > .ant-card-body {
-        padding: 0;
-      }
-    }
-  }
-}
-
-.footer {
-  display: flex;
-  justify-content: center;
-
-  color: lightgray;
-}
