+++ src/page/home/index.tsx
@@ -1,8 +1,7 @@
 import React from 'react';
-import Topics from '@/component/Topic';
 
 const Home: React.FC<Props> = () => {
-  return <Topics />;
+  return null;
 };
 
 export default Home;
